Fix permissions
Permissions for Flux GitRepository, OCIRepository and Bucket were
missing.
